added safeguards

+++ SolutionwebTest/SolutionwebTest/Controllers/HomeController.cs
@@ -26,6 +26,10 @@ namespace SolutionwebTest.Controllers
         [HttpPost]
         public IActionResult Index(FormInfo info)
         {
+            if (!ModelState.IsValid)
+            {
+                return View();
+            }
             _db.formInfos.Add(info);
             _db.SaveChanges();          
             return RedirectToAction("Update", "Home", new {id = info.Id });
@@ -55,15 +59,25 @@ namespace SolutionwebTest.Controllers
         }
 
         [HttpPost]
-        public IActionResult Delete(long id)
+        public IActionResult Delete(long? id)
         {
+            var forminfos = _db.formInfos.ToArray();
             Debug.WriteLine("inside delete trying to delete reccord with id: " + id);
+            if (id == null)
+            {
+                return View(forminfos);
+            }
             FormInfo existing = _db.formInfos.Find(id);
+            if (!(existing == null)) { 
             _db.formInfos.Remove(existing);
             _db.SaveChanges();
             ViewData["Message"] = "Delete table rows";
-            var forminfos = _db.formInfos.ToArray();
             return View(forminfos);
+            } 
+            ViewData["Message"] = "No row for id " + id;
+            return View();
+            
+        
         }
         public IActionResult Contact()
         {
